correccion menu de alta
